+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## [2.0.0] - 2025-02-28
+
+### Added
+
+- Package groups
+
+### Changed
+
+- Installing packages are now done by selecting them and then pressing the button "Install"
+
+### Fixed
+
 ## [1.0.0] - 2025-02-19
 
 ### Added